+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/jss/compare/v1.34.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- index to sitemap generator
 - sitemap change freq to config
 
 [1.34.0](https://github.com/jshwi/jss/releases/tag/v1.34.0) - 2023-12-04